requirements.txt
Django==3.2.1
django-admincharts==0.3.1
django-allauth==0.50.0
django-binary-database-files==1.0.15
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-coverage-plugin==2.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.8.0
django-material-admin==1.8.6
django-model-utils==4.2.0
django-modelcluster==5.2
django-modeltranslation==0.16.2
django-multiselectfield==0.1.12
django-soft-delete==0.9.21
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-suit==0.2.25
django-taggit==1.5.1
django-timezone-field==4.2.3
django-treebeard==4.5.1
djangocms-admin-style==3.1.1
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
l18n==2021.3
Markdown==3.3.4
PyJWT==1.7.1
pytest==6.2.5
requests==2.25.1
django-binary-database-files==1.0.15
Pillow==9.1.0
